# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Little Scholar Backend API",
    "version": "1.0.0",
    "description": "API specification for the Little Scholar backend. The service provides health checks, parent authentication, child profiles, AI exam generation, synced exam attempts, performance summaries, usage quotas, and AI learning insights for the Little Scholar iOS app.\n"
  },
  "servers": [
    {
      "url": "/",
      "description": "Same origin as the Swagger UI"
    },
    {
      "url": "https://your-railway-domain.up.railway.app",
      "description": "Railway deployment"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/children": {
      "post": {
        "tags": [
          "Children"
        ],
        "operationId": "createChild",
        "summary": "Create a child profile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChildProfileRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Child profile created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChildProfileResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/AuthError"
          }
        }
      },
      "get": {
        "tags": [
          "Children"
        ],
        "operationId": "listChildren",
        "summary": "List child profiles",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Child profiles for the authenticated parent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChildProfileListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/AuthError"
          }
        }
      }
    },
    "/api/children/{childId}": {
      "put": {
        "tags": [
          "Children"
        ],
        "operationId": "updateChild",
        "summary": "Update a child profile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ChildId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChildProfileUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Child profile updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChildProfileResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/AuthError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      },
      "delete": {
        "tags": [
          "Children"
        ],
        "operationId": "deleteChild",
        "summary": "Delete a child profile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ChildId"
          }
        ],
        "responses": {
          "200": {
            "description": "Child profile deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/AuthError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
    "/api/exam-attempts": {
      "post": {
        "tags": [
          "Exam Attempts"
        ],
        "operationId": "createExamAttempt",
        "summary": "Sync an exam attempt",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExamAttemptRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Exam attempt stored.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamAttemptResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/AuthError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
    "/api/exam-attempts/children/{childId}": {
      "get": {
        "tags": [
          "Exam Attempts"
        ],
        "operationId": "listExamAttemptsForChild",
        "summary": "List synced attempts for a child",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ChildId"
          }
        ],
        "responses": {
          "200": {
            "description": "Synced attempts for the child.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamAttemptListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/AuthError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Paste the JWT accessToken value only. Do not include the Bearer prefix."
      }
    },
    "parameters": {
      "ChildId": {
        "name": "childId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      }
    },
    "responses": {
      "ValidationError": {
        "description": "Request validation failed.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "validationFailed": {
                "value": {
                  "success": false,
                  "error": {
                    "message": "Validation failed",
                    "details": {
                      "body": {
                        "email": [
                          "Invalid email"
                        ]
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "AuthError": {
        "description": "Missing, invalid, or expired access token.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "authRequired": {
                "value": {
                  "success": false,
                  "error": {
                    "message": "Authentication required"
                  }
                }
              }
            }
          }
        }
      },
      "NotFoundError": {
        "description": "Requested resource was not found for the authenticated user.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "notFound": {
                "value": {
                  "success": false,
                  "error": {
                    "message": "Child profile not found"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "schemas": {
      "ApiSuccess": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "const": true
          },
          "data": {}
        },
        "additionalProperties": false
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "success",
          "error"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "const": false
          },
          "error": {
            "type": "object",
            "required": [
              "message"
            ],
            "properties": {
              "message": {
                "type": "string"
              },
              "details": {
                "description": "Optional validation or provider details."
              }
            },
            "additionalProperties": false
          }
        },
        "additionalProperties": false
      },
...
      "Grade": {
        "type": "string",
        "enum": [
          "Nursery",
          "LKG",
          "UKG",
          "Grade 1"
        ]
      },
      "Subject": {
        "type": "string",
        "enum": [
          "English",
          "Maths",
          "Hindi",
          "EVS",
          "GK"
        ]
      },
      "Difficulty": {
        "type": "string",
        "enum": [
          "Easy",
          "Medium",
          "Hard"
        ]
      },
...
      "ChildProfileRequest": {
        "type": "object",
        "required": [
          "name",
          "age",
          "grade"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80,
            "example": "Aarav"
          },
          "age": {
            "type": "integer",
            "minimum": 2,
            "maximum": 10,
            "example": 6
          },
          "grade": {
            "$ref": "#/components/schemas/Grade"
          },
          "avatarUrl": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri",
            "maxLength": 500
          }
        },
        "additionalProperties": false
      },
      "ChildProfileUpdateRequest": {
        "type": "object",
        "minProperties": 1,
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "age": {
            "type": "integer",
            "minimum": 2,
            "maximum": 10
          },
          "grade": {
            "$ref": "#/components/schemas/Grade"
          },
          "avatarUrl": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri",
            "maxLength": 500
          }
        },
        "additionalProperties": false
      },
      "ChildProfile": {
        "type": "object",
        "required": [
          "id",
          "userId",
          "name",
          "age",
          "grade",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "age": {
            "type": "integer"
          },
          "grade": {
            "$ref": "#/components/schemas/Grade"
          },
          "avatarUrl": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "ChildProfileResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiSuccess"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/ChildProfile"
              }
            }
          }
        ]
      },
      "ChildProfileListResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiSuccess"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ChildProfile"
                }
              }
            }
          }
        ]
      },
      "DeleteResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiSuccess"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "deleted"
                ],
                "properties": {
                  "deleted": {
                    "type": "boolean",
                    "const": true
                  }
                },
                "additionalProperties": false
              }
            }
          }
        ]
      },
      "CreateExamAttemptRequest": {
        "type": "object",
        "required": [
          "childId",
          "subject",
          "difficulty",
          "questionCount",
          "correctAnswers"
        ],
        "properties": {
          "childId": {
            "type": "string",
            "format": "uuid"
          },
          "examId": {
            "type": "string",
            "maxLength": 120
          },
          "grade": {
            "$ref": "#/components/schemas/Grade"
          },
          "subject": {
            "$ref": "#/components/schemas/Subject"
          },
          "difficulty": {
            "$ref": "#/components/schemas/Difficulty"
          },
          "questionCount": {
            "type": "integer",
            "minimum": 1,
            "maximum": 200
          },
          "correctAnswers": {
            "type": "integer",
            "minimum": 0,
            "maximum": 200
          },
          "totalMarks": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1000
          },
          "earnedMarks": {
            "type": "integer",
            "minimum": 0,
            "maximum": 1000
          },
          "scorePercentage": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "timeSpentSeconds": {
            "type": "integer",
            "minimum": 0,
            "maximum": 86400
          },
          "answers": {
            "description": "Optional client answer details."
          },
          "strongTopics": {
            "type": "array",
            "maxItems": 50,
            "items": {
              "type": "string"
            }
          },
          "weakTopics": {
            "type": "array",
            "maxItems": 50,
            "items": {
              "type": "string"
            }
          },
          "attemptedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "ExamAttempt": {
        "type": "object",
        "required": [
          "id",
          "userId",
          "childId",
          "grade",
          "subject",
          "difficulty",
          "questionCount",
          "correctAnswers",
          "scorePercentage",
          "strongTopics",
          "weakTopics",
          "attemptedAt",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "childId": {
            "type": "string",
            "format": "uuid"
          },
          "examId": {
            "type": [
              "string",
              "null"
            ]
          },
          "grade": {
            "$ref": "#/components/schemas/Grade"
          },
          "subject": {
            "$ref": "#/components/schemas/Subject"
          },
          "difficulty": {
            "$ref": "#/components/schemas/Difficulty"
          },
          "questionCount": {
            "type": "integer"
          },
          "correctAnswers": {
            "type": "integer"
          },
          "totalMarks": {
            "type": [
              "integer",
              "null"
            ]
          },
          "earnedMarks": {
            "type": [
              "integer",
              "null"
            ]
          },
          "scorePercentage": {
            "type": "number"
          },
          "timeSpentSeconds": {
            "type": [
              "integer",
              "null"
            ]
          },
          "answers": {
            "nullable": true
          },
          "strongTopics": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "weakTopics": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "attemptedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "ExamAttemptResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiSuccess"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/ExamAttempt"
              }
            }
          }
        ]
      },
      "ExamAttemptListResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiSuccess"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ExamAttempt"
                }
              }
            }
          }
        ]
      },
...
    }
  }
}